v2/requirements.txt
# Book Generator v2 dependencies
agno>=2.5.10
pydantic>=2.0
openai>=1.0
python-docx>=1.1.0
pyyaml>=6.0
